Update item image support to conform to the RSS spec.
Instead of an image section that is not allowed, this uses the enclosures as is expected. If image size and mime type are not passed, but image url is, we only use the itunes image tag. This can be desirable when we want the itunes tags, but no enclosures as some podcast apps show enclosures as downloads.